Moved view_file from utils/__init__.py to utils/fileutils.py
